LibUnicode: Move number formatting code generator to UnicodeNumberFormat
